Fix pylint and add extra test
