+++ src/.coveragerc
@@ -1,0 +1,21 @@
+[run]
+source = src
+branch = True
+omit =  */test*
+        */__init__.py*
+        */migrations/*
+        */manage.py*
+
+[report]
+exclude_lines =
+    pragma: no cover
+    def __repr__
+    if self\.debug
+    raise AssertionError
+    raise NotImplementedError
+    if 0:
+    if __name__ == .__main__.:
+ignore_errors = True
+
+[html]
+directory = htmlcov
