+++ src/members/admin.py
@@ -1,12 +1,8 @@
 """admin.py"""
 from django.contrib import admin
-from .models import Member
-from django.contrib import admin
-from rest_framework_api_key.admin import APIKeyModelAdmin
 from .models import APIKey
+from .models import Member
 
-@admin.register(APIKey)
-class APIKeyModelAdmin(APIKeyModelAdmin):
-    pass
 
+admin.site.register(APIKey)
 admin.site.register(Member)
