+++ src/members/test/tests.py
@@ -1,12 +1,8 @@
 
 """Test urls"""
-import json
-from django.test import SimpleTestCase, Client, TestCase, RequestFactory
-from rest_framework.test import force_authenticate
-from django.urls import reverse
+from django.test import SimpleTestCase
 from django.urls import resolve, reverse
 from ewilgs.views import *
-from ..models import Member
 
 # pylint: disable=all
 
@@ -15,33 +11,5 @@ class TestUrls(SimpleTestCase):
         url = reverse('home')
         self.assertEquals(resolve(url).func, home)
 
-class TestSubmitFrames(TestCase):
-    def setUp(self):
-        self.factory = RequestFactory()
-        self.client = Client()
-        self.user = Member.objects.create_user(username='user', email='user@email.com')
-        self.user.set_password('delfispace4242')
-        self.user.save()
 
-    def tearDown(self):
-        self.client.logout()
-
-    def test_submit(self):
-
-        self.assertEqual(len(Downlink.objects.all()), 0) # downlink table empty
-
-        response = self.client.post(reverse('login'), {'username': 'user', 'password': 'delfispace4242'})
-        self.assertEqual(response.status_code, 200)
-        frame = '{ "qos": 98.6, "pass": "qwerty", "frame_time": "2021-12-19T02:20:14.959630Z", "frequency": 2455.66, "frame": "A8989A40404000888C9C66B0A80003F0890FFDAD776500001E601983C008C39C10D02911E2F0FF71230DECE70032044C09500311119B8CA092A08B5E85919492938285939C7900000000000000000000005602F637005601F3380000006D70000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000434B1345B440BF3C9736D0301D240E000004B82C4050B26DDB942EB4D0CFE4E9D64946"}'
-
-        frame_json = json.loads(frame)
-        request = self.factory.post(path='submit', data=frame_json, content_type='application/json')
-
-        user = Member.objects.get(username='user')
-        force_authenticate(request, user=user)
-        request.user = user
-        response = submit(request)
-
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(len(Downlink.objects.all()), 1) # dowlink table has 1 entry
